refactor: ordertable 패키지를 eatinorders 패키지 하위에 생성
