+++ src/main/java/kitchenpos/eatinorders/ordertable/domain/OrderTable.java
@@ -1,4 +1,4 @@
-package kitchenpos.eatinorders.domain.ordertable;
+package kitchenpos.eatinorders.ordertable.domain;
 
 import java.util.UUID;
