+++ src/main/java/kitchenpos/eatinorders/ordertable/infra/JpaOrderTableRepository.java
@@ -1,11 +1,11 @@
-package kitchenpos.eatinorders.infra;
+package kitchenpos.eatinorders.ordertable.infra;
 
 import java.util.UUID;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import kitchenpos.eatinorders.domain.ordertable.OrderTable;
-import kitchenpos.eatinorders.domain.ordertable.OrderTableRepository;
+import kitchenpos.eatinorders.ordertable.domain.OrderTable;
+import kitchenpos.eatinorders.ordertable.domain.OrderTableRepository;
 
 public interface JpaOrderTableRepository extends OrderTableRepository, JpaRepository<OrderTable, UUID> {
 }
